Extracing makeExampleImplementation.
