+++ src/integration.ts
@@ -238,160 +238,7 @@ export class Integration extends IntegrationBase {
       this.makeClientClass(),
     );
 
-    // method: method.toUpperCase()
-    const methodProperty = f.createPropertyAssignment(
-      this.ids.methodParameter,
-      makePropCall(this.ids.methodParameter, propOf<string>("toUpperCase")),
-    );
-
-    // headers: hasBody ? { "Content-Type": "application/json" } : undefined
-    const headersProperty = f.createPropertyAssignment(
-      this.ids.headersProperty,
-      makeTernary(
-        this.ids.hasBodyConst,
-        f.createObjectLiteralExpression([
-          f.createPropertyAssignment(
-            f.createStringLiteral("Content-Type"),
-            f.createStringLiteral(contentTypes.json),
-          ),
-        ]),
-        this.ids.undefinedValue,
-      ),
-    );
-
-    // body: hasBody ? JSON.stringify(params) : undefined
-    const bodyProperty = f.createPropertyAssignment(
-      this.ids.bodyProperty,
-      makeTernary(
-        this.ids.hasBodyConst,
-        makePropCall(f.createIdentifier("JSON"), propOf<JSON>("stringify"), [
-          this.ids.paramsArgument,
-        ]),
-        this.ids.undefinedValue,
-      ),
-    );
-
-    // const response = await fetch(new URL(`${path}${searchParams}`, "https://example.com"), { ___ });
-    const responseStatement = makeConst(
-      this.ids.responseConst,
-      f.createAwaitExpression(
-        f.createCallExpression(f.createIdentifier(fetch.name), undefined, [
-          makeNew(
-            f.createIdentifier(URL.name),
-            makeTemplate(
-              "",
-              [this.ids.pathParameter],
-              [this.ids.searchParamsConst],
-            ),
-            f.createStringLiteral(serverUrl),
-          ),
-          f.createObjectLiteralExpression([
-            methodProperty,
-            headersProperty,
-            bodyProperty,
-          ]),
-        ]),
-      ),
-    );
-
-    // const hasBody = !["get", "delete"].includes(method);
-    const hasBodyStatement = makeConst(
-      this.ids.hasBodyConst,
-      f.createLogicalNot(
-        makePropCall(
-          f.createArrayLiteralExpression([
-            f.createStringLiteral("get" satisfies Method),
-            f.createStringLiteral("delete" satisfies Method),
-          ]),
-          propOf<string[]>("includes"),
-          [this.ids.methodParameter],
-        ),
-      ),
-    );
-
-    // const searchParams = hasBody ? "" : `?${new URLSearchParams(params)}`;
-    const searchParamsStatement = makeConst(
-      this.ids.searchParamsConst,
-      makeTernary(
-        this.ids.hasBodyConst,
-        f.createStringLiteral(""),
-        makeTemplate("?", [
-          makeNew(
-            f.createIdentifier(URLSearchParams.name),
-            this.ids.paramsArgument,
-          ),
-        ]),
-      ),
-    );
-
-    // const contentType = response.headers.get("content-type");
-    const contentTypeStatement = makeConst(
-      this.ids.contentTypeConst,
-      makePropCall(
-        [this.ids.responseConst, this.ids.headersProperty],
-        propOf<Headers>("get"),
-        [f.createStringLiteral("content-type")],
-      ),
-    );
-
-    // if (!contentType) return;
-    const noBodyStatement = f.createIfStatement(
-      f.createPrefixUnaryExpression(
-        ts.SyntaxKind.ExclamationToken,
-        this.ids.contentTypeConst,
-      ),
-      f.createReturnStatement(),
-    );
-
-    // const isJSON = contentType.startsWith("application/json");
-    const isJsonConst = makeConst(
-      this.ids.isJsonConst,
-      makePropCall(this.ids.contentTypeConst, propOf<string>("startsWith"), [
-        f.createStringLiteral(contentTypes.json),
-      ]),
-    );
-
-    // return response[isJSON ? "json" : "text"]();
-    const returnStatement = f.createReturnStatement(
-      f.createCallExpression(
-        f.createElementAccessExpression(
-          this.ids.responseConst,
-          makeTernary(
-            this.ids.isJsonConst,
-            f.createStringLiteral(propOf<Response>("json")),
-            f.createStringLiteral(propOf<Response>("text")),
-          ),
-        ),
-        undefined,
-        [],
-      ),
-    );
-
-    // export const exampleImplementation: Implementation = async (method,path,params) => { ___ };
-    const exampleImplStatement = makeConst(
-      this.ids.exampleImplementationConst,
-      makeArrowFn(
-        [
-          this.ids.methodParameter,
-          this.ids.pathParameter,
-          this.ids.paramsArgument,
-        ],
-        f.createBlock([
-          hasBodyStatement,
-          searchParamsStatement,
-          responseStatement,
-          contentTypeStatement,
-          noBodyStatement,
-          isJsonConst,
-          returnStatement,
-        ]),
-        { isAsync: true },
-      ),
-      {
-        expose: true,
-        type: f.createTypeReferenceNode(this.ids.implementationType),
-      },
-    );
+    const exampleImplStatement = this.makeExampleImplementation(serverUrl);
 
     // client.provide("get /v1/user/retrieve", { id: "10" });
     const provideCallingStatement = f.createExpressionStatement(
